+++ bin/my_clean_cli.dart
@@ -1,9 +1,18 @@
+
 import 'dart:io';
 import 'package:args/args.dart';
+import 'package:path/path.dart' as path;
+import 'package:http/http.dart' as http;
+import 'package:archive/archive.dart';
 import 'package:yaml/yaml.dart';
 import 'package:yaml_edit/yaml_edit.dart';
 
-
+/// The main function of the CLI tool.
+/// It parses the command-line arguments and calls the appropriate functions.
+/// The CLI tool supports two commands: create and update-pubspec.
+/// The create command downloads a template ZIP file from GitHub and extracts the contents.
+/// The update-pubspec command adds predefined dependencies and assets to the pubspec.yaml file.
+/// The predefined dependencies are defined in the function.
 void main(List<String> arguments) {
   final parser = ArgParser()
     ..addFlag('create', abbr: 'c', help: 'Create predefined folders and files')
@@ -19,6 +28,7 @@ void main(List<String> arguments) {
     updatePubspec();
   }
 }
+/// createAsset function is used to create the assets folder in the root directory of the project
 void createAsset() {
   // Define the root-level folders to be created
   const folders = [
@@ -38,38 +48,49 @@ void createAsset() {
     }
   }
 }
-void createFoldersAndFiles() {
-  final templateDir = Directory('template');
-  final targetDir = Directory('lib');
-createAsset();
-  if (!templateDir.existsSync()) {
-    print('Template directory not found!');
-    return;
-  }
 
-  // Copy files and directories from template to the target directory
-  templateDir.listSync(recursive: true).forEach((entity) {
-    if (entity is Directory) {
-      final relativePath = entity.path.replaceFirst(templateDir.path, '');
-      final targetSubDir = Directory(targetDir.path + relativePath);
-      if (!targetSubDir.existsSync()) {
-        targetSubDir.createSync(recursive: true);
-        print('Created directory: ${targetSubDir.path}');
+/// createFoldersAndFiles function is used to create the predefined folders and files in the project
+/// by downloading a template ZIP file from GitHub and extracting the contents.
+/// The template contains the predefined folder structure and some files.
+Future<void> createFoldersAndFiles() async {
+  final targetDir = Directory('lib/');
+  createAsset();
+
+  try {
+    // Download the template ZIP file from GitHub
+    final url = 'https://github.com/farooq958/cli_template/archive/refs/heads/master.zip';
+    final response = await http.get(Uri.parse(url));
+    final archive = ZipDecoder().decodeBytes(response.bodyBytes);
+
+    // Extract only the files and directories from lib/template/
+    final templatePrefix = 'cli_template-master/lib/template/';
+
+    for (final file in archive) {
+      if (file.name.startsWith(templatePrefix)) {
+        final relativePath = file.name.substring(templatePrefix.length);
+        final filePath = path.join(targetDir.path, relativePath);
+
+        if (file.isFile) {
+          final outputFile = File(filePath);
+          outputFile.createSync(recursive: true);
+          outputFile.writeAsBytesSync(file.content as List<int>);
+          print('Created file: $filePath');
+        } else if (file is Directory) {
+          final outputDir = Directory(filePath);
+          outputDir.createSync(recursive: true);
+          print('Created directory: $filePath');
+        }
       }
-    } else if (entity is File) {
-      final relativePath = entity.path.replaceFirst(templateDir.path, '');
-      final targetFile = File(targetDir.path + relativePath);
-
-      // Ensure target directory exists
-      targetFile.parent.createSync(recursive: true);
-
-      // Copy file as is
-      entity.copySync(targetFile.path);
-      print('Created file: ${targetFile.path}');
     }
-  });
+  } catch (e) {
+    print('Error while creating files: $e');
+  }
 }
 
+/// updatePubspec function is used to add predefined dependencies and assets to the pubspec.yaml file
+/// if they don't already exist.
+/// It also creates the flutter section if it doesn't exist.
+/// The predefined dependencies are defined in the function.
 void updatePubspec() {
   final pubspecFile = File('pubspec.yaml');
   if (!pubspecFile.existsSync()) {
@@ -92,17 +113,16 @@ void updatePubspec() {
     'flutter_launcher_icons': '^0.13.1',
     'firebase_core': '^3.2.0',
     'firebase_auth': '^5.1.2',
-    'firebase_messaging': '^15.0.3',
+    'firebase_messaging': '^15.0.4',
     'flutter_bloc': '^8.1.2',
     'flutter_svg': '^2.0.10+1',
-    'pinput': '^5.0.0',
     'carousel_slider_plus': '^7.0.0',
     'connectivity_plus': '^6.0.3',
     'permission_handler': '^11.3.1',
     'path_provider': '^2.0.11',
     'dio': '^5.5.0+1',
     'flutter_local_notifications': '^17.1.0',
-    'share_plus': '^9.0.0',
+    'share_plus': '^10.0.0',
     'dio_cache_interceptor': '^3.5.0',
     'flutter_quick_router': '^0.0.2',
     'country_picker': '^2.0.26',
@@ -111,14 +131,10 @@ void updatePubspec() {
     'quick_router': '^1.0.1',
     'gif_view': '^0.4.3',
     'dotted_border': '^2.1.0',
-    'syncfusion_flutter_datepicker': '^26.1.41',
     'flutter_animate': '^4.5.0',
     'google_fonts': 'any',
     'video_player': '^2.9.1',
-    'google_maps_flutter': '^2.7.0',
     'lottie': '^3.1.2',
-    'share_it': '^0.7.0',
-    'flutter_contacts': '^1.1.9',
     'image_picker': '^0.8.4+4',
     'image_cropper': '^1.4.0',
     'grouped_list': 'any',
@@ -173,3 +189,4 @@ void updatePubspec() {
   pubspecFile.writeAsStringSync(yamlEditor.toString());
   print('pubspec.yaml updated successfully.');
 }
+