+++ template/data/app_data/prefs_keys.dart
@@ -1,9 +1,0 @@
-class PrefsKeys{
-
-  static String USER_KEY= 'user';
-  static String TOKEN_KEY= 'token';
-  static String THEME_KEY="themeMode";
-  static  String LOCALE_KEY='localeKey';
-
-
-}