+++ template/data/app_data/user_data.dart
@@ -1,5 +1,0 @@
-mixin UserData {
-  static String? userToken;
-  set token(String? token) => userToken = token;
-  String? get token => userToken;
-}