+++ template/presentation/common/quick_frame/controller/btController.dart
@@ -1,8 +1,0 @@
-import 'package:flutter/material.dart';
-
-
-class BottomNotifier {
-  static ValueNotifier<int> bottomNavigationNotifier = ValueNotifier(0);
-  static PageController? bottomPageController = PageController(initialPage: 0);
-
-}