+++ CHANGELOG.md
@@ -9,3 +9,13 @@
 - Initial release of `clean_architecture_cli`.
 - Added functionality to generate predefined folder structures.
 - Added functionality to update the `pubspec.yaml` file with dependencies and assets.
+
+## 1.0.6    
+- Minor fix but important for the user
+
+## 1.0.7
+- Added functionality to generate predefined widgets.
+
+## 1.0.8
+- Updated functionality to generate predefined widgets.
+- Updated functionality to generate pubspec dependency.